+++ scripts/higgs.py
@@ -13,7 +13,25 @@ import numpy as np
 import pandas as pd
 
 from utils import column_name_array, default_config, generate_profile, save_json, save_npz, \
-    split_by_label
+    split_by_label, uci_license
+
+bibtex = """\
+@article{baldiSearchingExoticParticles2014,
+  title = {Searching for Exotic Particles in High-Energy Physics with Deep Learning},
+  author = {Baldi, P. and Sadowski, P. and Whiteson, D.},
+  date = {2014-07-02},
+  journaltitle = {Nature Communications},
+  shortjournal = {Nat Commun},
+  volume = {5},
+  number = {1},
+  pages = {4308},
+  publisher = {{Nature Publishing Group}},
+  issn = {2041-1723},
+  doi = {10.1038/ncomms5308},
+  url = {https://www.nature.com/articles/ncomms5308},
+  urldate = {2021-12-13},
+  issue = {1},
+}"""
 
 column_names = [
     'label',
@@ -56,17 +74,18 @@ def convert_format(config):
         index_col=None,
         names=column_names
     )
+
+    # split into their standard train/test sets
+    train_df = df[:-500_000]
+    test_df = df[-500_000:]
     
     if config.dataset_file:
-        # split into their standard train/test sets
-        train_df = df[:-500_000]
-        test_df = df[-500_000:]
-        
         train_data_df, train_labels_df = split_by_label(train_df)
         test_data_df, test_labels_df = split_by_label(test_df)
         
         save_npz(
-            os.path.join(config.outputdirectory, 'higgs'), {
+            config,
+            {
                 'train-data': train_data_df,
                 'train-labels': train_labels_df,
                 'test-data': test_data_df,
@@ -75,15 +94,27 @@ def convert_format(config):
                 '_columns-labels': np.array(['label'], dtype=np.str_),
             }
         )
-
+    
     if config.extras_file:
-        save_json({
-            'profile': generate_profile(df),
-        }, os.path.join(config.outputdirectory, 'higgs.json'))
+        save_json(
+            config,
+            {
+                'train-profile': generate_profile(train_df, config.no_profile),
+                'profile': generate_profile(df, config.no_profile),
+                'bibtex': bibtex,
+                'license': uci_license,
+                'categories': {
+                    'label': ['background', 'signal'],
+                },
+                'column-names-attributes': column_names[1:],
+                'column-names-target': column_names[:1],
+            }
+        )
 
 
 if __name__ == '__main__':
     config = default_config(
+        'higgs',
         download_root='https://archive.ics.uci.edu/ml/machine-learning-databases/00280/',
     )
     convert_format(config)