+++ scripts/musk.py
@@ -6,8 +6,9 @@ import os
 
 from sklearn.model_selection import train_test_split
 
-from utils import column_name_array, create_csv_reader, default_config, save_npz, \
-    split_by_label
+from utils import column_name_array, create_csv_reader, default_config, generate_profile, save_json, \
+    save_npz, \
+    split_by_label, uci_bibtex, uci_license
 
 column_names = [
     'molecule_name',  # NOT used as input
@@ -44,20 +45,36 @@ def convert_format(config):
     train_data_df, train_labels_df = split_by_label(train_df, 'musk')
     test_data_df, test_labels_df = split_by_label(test_df, 'musk')
     
-    save_npz(
-        os.path.join(config.outputdirectory, 'musk'), {
-            'train-data': train_data_df,
-            'train-labels': train_labels_df,
-            'test-data': test_data_df,
-            'test-labels': test_labels_df,
-            '_columns-data': column_name_array(train_data_df),
-            '_columns-labels': column_name_array(train_labels_df),
-        }
-    )
+    if config.dataset_file:
+        save_npz(
+            config,
+            {
+                'train-data': train_data_df,
+                'train-labels': train_labels_df,
+                'test-data': test_data_df,
+                'test-labels': test_labels_df,
+                '_columns-data': column_name_array(train_data_df),
+                '_columns-labels': column_name_array(train_labels_df),
+            }
+        )
+    
+    if config.extras_file:
+        save_json(
+            config,
+            {
+                'train-profile': generate_profile(train_df, config.no_profile),
+                'profile': generate_profile(df, config.no_profile),
+                'bibtex': uci_bibtex,
+                'license': uci_license,
+                'column-names-attributes': list(train_data_df.columns),
+                'column-names-target': list(train_labels_df.columns),
+            }
+        )
 
 
 if __name__ == '__main__':
     args = default_config(
+        'musk',
         download_root='https://archive.ics.uci.edu/ml/machine-learning-databases/musk/',
     )