+++ scripts/poker.py
@@ -5,8 +5,9 @@ its train-test split.
 
 import os
 
-from utils import column_name_array, create_csv_reader, default_config, save_npz, \
-    split_by_label
+from utils import column_name_array, create_csv_reader, default_config, generate_profile, hvcat, save_json, \
+    save_npz, \
+    split_by_label, uci_bibtex, uci_license
 
 column_names = [
     'S1',
@@ -19,8 +20,18 @@ column_names = [
     'C4',
     'S5',
     'C5',
-    'label',
+    'hand',
 ]
+
+suits = ['hearts', 'spades', 'diamonds', 'clubs']
+ranks = ['ace', '2', '3', '4', '5', '6', '7', '8', '9', '10', 'jack', 'queen', 'king']
+
+categories = {
+    'hand': ['nothing', 'one pair', 'two pairs', 'three of a kind', 'straight',
+             'flush', 'full house', 'four of a kind', 'straight flush', 'royal flush'],
+    **{f'S{c}': suits for c in range(1, 5+1)},
+    **{f'C{c}': ranks for c in range(1, 5+1)},
+}
 
 
 def convert_format(config):
@@ -34,23 +45,43 @@ def convert_format(config):
     train_df = read_csv('poker-hand-training-true.data')
     test_df = read_csv('poker-hand-testing.data')
     
-    train_data_df, train_labels_df = split_by_label(train_df)
-    test_data_df, test_labels_df = split_by_label(test_df)
+    train_df[column_names[:-1]] = train_df[column_names[:-1]] - 1
+    test_df[column_names[:-1]] = test_df[column_names[:-1]] - 1
     
-    save_npz(
-        os.path.join(config.outputdirectory, 'poker'), {
-            'train-data': train_data_df,
-            'train-labels': train_labels_df,
-            'test-data': test_data_df,
-            'test-labels': test_labels_df,
-            '_columns-data': column_name_array(train_data_df),
-            '_columns-labels': column_name_array(train_labels_df),
-        }
-    )
+    train_data_df, train_labels_df = split_by_label(train_df, 'hand')
+    test_data_df, test_labels_df = split_by_label(test_df, 'hand')
+    
+    if config.dataset_file:
+        save_npz(
+            config,
+            {
+                'train-data': train_data_df,
+                'train-labels': train_labels_df,
+                'test-data': test_data_df,
+                'test-labels': test_labels_df,
+                '_columns-data': column_name_array(train_data_df),
+                '_columns-labels': column_name_array(train_labels_df),
+            }
+        )
+    
+    if config.extras_file:
+        save_json(
+            config,
+            {
+                'train-profile': generate_profile(train_df, config.no_profile),
+                'profile': generate_profile(hvcat([[train_df], [test_df]]), config.no_profile),
+                'categories': categories,
+                'license': uci_license,
+                'bibtex': uci_bibtex,
+                'column-names-attributes': column_names[:-1],
+                'column-names-target': column_names[-1:],
+            }
+        )
 
 
 if __name__ == '__main__':
     args = default_config(
+        'poker',
         download_root='https://archive.ics.uci.edu/ml/machine-learning-databases/poker',
     )