+++ scripts/rossman.py
@@ -12,7 +12,7 @@ import simplejson as json
 from sklearn.model_selection import train_test_split
 
 from utils import column_name_array, convert_categorical, create_csv_reader, default_config, \
-    save_npz, split_by_label
+    generate_profile, save_json, save_npz, split_by_label
 
 state_holidays = {
     'a': 'public holiday',
@@ -49,7 +49,7 @@ def convert_format(config):
     store_df['PromoInterval'] = store_df['PromoInterval'].astype('category')
     
     df = train_df.join(store_df.set_index('Store'), 'Store')
-    categories = convert_categorical(df)
+    df, categories = convert_categorical(df)
     
     # split between stores (stores only show up in one split)
     train_stores, test_stores = train_test_split(
@@ -63,25 +63,32 @@ def convert_format(config):
     train_data_df, train_labels_df = split_by_label(train_df, 'Sales')
     test_data_df, test_labels_df = split_by_label(test_df, 'Sales')
     
-    save_npz(
-        os.path.join(config.outputdirectory, 'rossman'), {
-            'train-data': train_data_df,
-            'train-labels': train_labels_df,
-            'test-data': test_data_df,
-            'test-labels': test_labels_df,
-            '_columns-data': column_name_array(train_data_df),
-            '_columns-labels': column_name_array(train_labels_df),
-        }
-    )
-    
-    with open(os.path.join(config.outputdirectory, 'rossman.json'), 'w') as f:
-        json.dump(
+    if config.dataset_file:
+        save_npz(
+            config,
             {
+                'train-data': train_data_df,
+                'train-labels': train_labels_df,
+                'test-data': test_data_df,
+                'test-labels': test_labels_df,
+                '_columns-data': column_name_array(train_data_df),
+                '_columns-labels': column_name_array(train_labels_df),
+            }
+        )
+    
+    if config.extras_file:
+        save_json(
+            config,
+            {
+                'train-profile': generate_profile(train_df, config.no_profile),
+                'profile': generate_profile(df, config.no_profile),
                 'categories': categories,
-            }, f
+                'column-names-attributes': list(train_data_df.columns),
+                'column-names-target': list(train_labels_df.columns),
+            }
         )
 
 
 if __name__ == '__main__':
-    args = default_config()
+    args = default_config('rossman')
     convert_format(args)