+++ scripts/utils.py
@@ -10,12 +10,14 @@ from pathlib import Path
 import numpy as np
 import pandas as pd
 import tarfile
-from dataprofiler import Profiler
+from dataprofiler import Profiler, ProfilerOptions
 
 try:
     import simplejson as json
 except ImportError:
     import json
+
+uci_license = 'CC BY 4.0'
 
 uci_bibtex = """\
 @misc{Dua:2019 ,
@@ -41,7 +43,7 @@ def default_config(name, *,
     parser = argparse.ArgumentParser(description=description)
     
     parser.add_argument(
-        'outputdirectory', dest='output_directory', type=Path,
+        'output_directory', type=Path,
         help='local directory to save outputs'
     )
     parser.add_argument(
@@ -92,7 +94,7 @@ def hvcat(arrays):
     )
 
 
-def save_npz(config, df_dict):
+def save_npz(config, df_dict, *, version=(1, 1, 0)):
     filename = (config.output_directory / config.name).with_suffix('.npz')
     comp_filename = filename.with_suffix('.npz.tar.gz')
     
@@ -100,7 +102,7 @@ def save_npz(config, df_dict):
         key: value.to_numpy().squeeze() if not isinstance(value, np.ndarray) else value
         for key, value in df_dict.items()
     }
-    arr_dict['_version'] = '1.0.0'
+    arr_dict['_version'] = np.array(version, dtype=np.int_)
     
     np.savez_compressed(str(filename), **arr_dict)
     print(f'\u001b[32mData saved in NPZ format at `{filename}`\u001b[0m')
@@ -131,18 +133,30 @@ def save_json(config, data):
     comp_filename = filename.with_suffix('.json.tar.gz')
     
     with filename.open('w') as f:
-        json.dump(data, f, cls=JSONNumpyEncoder)
-    
+        json.dump(
+            {
+                key: value for key, value in data.items() if value is not None
+            }, f, cls=JSONNumpyEncoder
+        )
     print(f'\u001b[32mExtras saved in JSON format at `{filename}`\u001b[0m')
     
     with tarfile.open(comp_filename, 'w:gz') as f:
         f.add(filename, arcname=filename.name)
-        
+    
     print(f'\u001b[32mExtras tarred and gzipped at `{comp_filename}`\u001b[0m')
 
 
-def generate_profile(df):
-    profile = Profiler(df)
+def generate_profile(df, no_profile=False, options=None):
+    if no_profile:
+        return None
+    
+    profile_options = ProfilerOptions()
+    profile_options.structured_options.data_labeler.is_enabled = False
+    profile_options.structured_options.chi2_homogeneity.is_enabled = False
+    if options is not None:
+        profile_options.set(options)
+    
+    profile = Profiler(df, options=profile_options)
     return profile.report(report_options={"output_format": "serializable"})
 
 
@@ -154,9 +168,10 @@ def convert_categorical(df):
     for column in categorical_columns:
         categories[column] = list(df[column].cat.categories)
     
-    df[categorical_columns] = df[categorical_columns].apply(lambda x: x.cat.codes)
+    cat_df = df.copy()
+    cat_df[categorical_columns] = df[categorical_columns].apply(lambda x: x.cat.codes)
     
-    return categories
+    return cat_df, categories
 
 
 @contextlib.contextmanager
@@ -164,4 +179,3 @@ def stage(action):
     print(f'Begin {action}...', end='', flush=True)
     yield
     print('finished.')
-